Waives some failing tests
