+++ tripy/tests/integration/conftest.py
@@ -1,5 +1,5 @@
 #
-# SPDX-FileCopyrightText: Copyright (c) 2024-2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2025-2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -58,4 +58,5 @@ def eager_or_compiled(request):
 
         return compiled_func(*tensor_args, **tensor_kwargs)
 
+    wrapper.mode = request.param
     return wrapper
